+++ pkg.nix
@@ -1,12 +1,12 @@
-{ mkDerivation, base, bytestring, network-simple-tls
+{ mkDerivation, async, base, bytestring, network-simple-tls
 , network-simple-ws, safe-exceptions, stdenv, websockets
 }:
 mkDerivation {
   pname = "network-simple-wss";
-  version = "0.1";
+  version = "0.1.1";
   src = ./.;
   libraryHaskellDepends = [
-    base bytestring network-simple-tls network-simple-ws
+    async base bytestring network-simple-tls network-simple-ws
     safe-exceptions websockets
   ];
   homepage = "https://github.com/k0001/network-simple-wss";