+++ changelog.md
@@ -1,3 +1,10 @@
+# Version 0.1.1
+
+* Depend on `network-simple-tls >=0.4`.
+
+* Depend on `websockets >=0.12.6`.
+
+
 # Version 0.1
 
 * First release.